--- src/PRISM/__init__.py
@@ -1,3 +1,0 @@
-from .PRISM_Data import *
-from .Percent_Precip_as_Snow import *
-from .Seasonality_Index import *
