+++ src/chmdata/__init__.py
@@ -1,3 +1,4 @@
 from .thredds import *
 from .mesonet import *
 from .agrimet import *
+from .prism import *
